Moved kfold functions from train to experiment
